requirements.txt
openai
sounddevice
scipy
playsound
python-dotenv
streamlit
soundfile
